MED-117 chore: rename linkHome to link
