+++ apps/medon-fe/src/components/Link/index.tsx
@@ -1,0 +1,6 @@
+import { LinkHomeProps } from 'components/Link/types';
+import { StyledLink } from 'components/Link/styles';
+
+export default function LinkHome({ children, ...rest }: LinkHomeProps) {
+  return <StyledLink {...rest}>{children}</StyledLink>;
+}
