+++ apps/medon-fe/src/components/__tests__/LinkHome.spec.tsx
@@ -1,14 +1,14 @@
 import { render, screen } from '@testing-library/react';
-import LinkHome from 'components/LinkHome';
+import Link from 'components/Link';
 import { TestWrapper } from 'utils/tests/TestWrapper';
 
-describe('LinkHome component', () => {
+describe('Link component', () => {
   it('renders a link with the provided text color and background color', () => {
     render(
       <TestWrapper>
-        <LinkHome to="/" textcolor="#fff" bgcolor="#000">
+        <Link to="/" textcolor="#fff" bgcolor="#000">
           Home
-        </LinkHome>
+        </Link>
       </TestWrapper>
     );
     const link = screen.getByRole('link');
@@ -20,9 +20,9 @@ describe('LinkHome component', () => {
   it('renders a full-width link when isfullwidth prop is true', () => {
     render(
       <TestWrapper>
-        <LinkHome to="/" isfullwidth="true">
+        <Link to="/" isfullwidth="true">
           Home
-        </LinkHome>
+        </Link>
       </TestWrapper>
     );
     const link = screen.getByRole('link');
@@ -33,7 +33,7 @@ describe('LinkHome component', () => {
   it('renders a link with the provided to prop', () => {
     render(
       <TestWrapper>
-        <LinkHome to="/">Home</LinkHome>
+        <Link to="/">Home</Link>
       </TestWrapper>
     );
     const link = screen.getByRole('link');
