+++ apps/medon-fe/src/pages/ResetPassword/index.tsx
@@ -1,12 +1,12 @@
 import { useTheme } from 'styled-components';
 import { useTranslation } from 'react-i18next';
-import { Link, useParams } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { useState } from 'react';
 
 import Button from 'components/Button';
-import LinkHome from 'components/LinkHome';
+import Link from 'components/Link';
 
 import RightArrow from 'assets/svgs/arrow/right-arrow.svg';
 import Logo from 'assets/svgs/logo_medon.svg';
@@ -98,18 +98,30 @@ export default function ResetPassword() {
             </>
           )}
         </Form>
-        <LinkHome
+        <Link
           bgcolor={theme.colors.black}
           textcolor={theme.colors.white}
           to="/"
           isfullwidth="true"
         >
           {t('forget-password.send-email.home-link')}
-        </LinkHome>
+        </Link>
       </Content>
       <Footer>
-        <Link to="#">{t('forget-password.footer.linkTerm')}</Link>
-        <Link to="#">{t('forget-password.footer.linkPrivacy')}</Link>
+        <Link
+          bgcolor={theme.colors.transparent}
+          textcolor={theme.colors.blue_300}
+          to="#"
+        >
+          {t('forget-password.footer.linkTerm')}
+        </Link>
+        <Link
+          bgcolor={theme.colors.transparent}
+          textcolor={theme.colors.blue_300}
+          to="#"
+        >
+          {t('forget-password.footer.linkPrivacy')}
+        </Link>
       </Footer>
     </Container>
   );
